refactor(entries): fix canceManuallSeeding filename spelling
Rename the misspelled canceManuallSeeding.ts to cancelManualSeeding.ts
(matching its already-correct cancelManualSeeding export) and update the
two importers. No behavior change.
